rename: chimei-ruiju: `service.rs` -> `chimei_ruiju.rs`
