+++ repository/src/chimei_ruiju/town.rs
@@ -1,7 +1,7 @@
 use domain::entity::TownMaster;
 use domain::error::ApiError;
 use jisx0401::Prefecture;
-use service::service::ChimeiRuijuApiService;
+use service::chimei_ruiju::ChimeiRuijuApiService;
 
 pub struct TownMasterRepository {
     api_service: ChimeiRuijuApiService,
@@ -28,7 +28,7 @@ impl TownMasterRepository {
 mod async_tests {
     use crate::chimei_ruiju::town::TownMasterRepository;
     use jisx0401::Prefecture;
-    use service::service::ChimeiRuijuApiService;
+    use service::chimei_ruiju::ChimeiRuijuApiService;
 
     #[tokio::test]
     async fn 東京都千代田区千代田() {
@@ -65,7 +65,7 @@ impl TownMasterRepository {
 mod blocking_tests {
     use crate::chimei_ruiju::town::TownMasterRepository;
     use jisx0401::Prefecture;
-    use service::service::ChimeiRuijuApiService;
+    use service::chimei_ruiju::ChimeiRuijuApiService;
 
     #[test]
     fn 京都府京都市伏見区魚屋町() {
