+++ service/src/lib.rs
@@ -1,1 +1,1 @@
-pub mod service;
+pub mod chimei_ruiju;
